fix minor bugs related with dlrmv3 kuairand reproduction (#277)
Summary:
Hi team,

Thank you for the excellent work on releasing DLRMv3 as a reference for generative recommenders!

While reproducing the results on the Kuairand dataset, I encountered a few issues that I would like to address in this PR. Could you please help review the changes?

1. The model is not being updated correctly.
2. The evaluation metrics are not calculated properly.
3. Some minor fixes to improve code usability.

Pull Request resolved: https://github.com/meta-recsys/generative-recommenders/pull/277

Reviewed By: ruochen99

Differential Revision: D74752474

Pulled By: LinjianMa

fbshipit-source-id: 4cd580390034782b3299ad9f7e4e01d138ea75a4